requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
decorator==5.1.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
rich==13.9.1
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
starlette==0.38.6
tornado==6.4.1
trio==0.26.2
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
